Remove deprecated layers and rename ChunkedImageLayer -> ImageLayer
