+++ examples/chunk_streaming/chunk_info_overlay.ts
@@ -1,15 +1,15 @@
 import { Idetik } from "../../src/idetik";
-import { ChunkedImageLayer } from "../../src/layers/chunked_image_layer";
+import { ImageLayer } from "../../src/layers/image_layer";
 import { Chunk } from "../../src/data/chunk";
 
 export interface ChunkInfoOverlayOptions {
   textDiv: HTMLDivElement;
-  imageLayer: ChunkedImageLayer;
+  imageLayer: ImageLayer;
 }
 
 export class ChunkInfoOverlay {
   private readonly textDiv_: HTMLDivElement;
-  private readonly imageLayer_: ChunkedImageLayer;
+  private readonly imageLayer_: ImageLayer;
 
   constructor({ textDiv, imageLayer }: ChunkInfoOverlayOptions) {
     this.textDiv_ = textDiv;
