+++ examples/chunk_streaming/scale_bar.ts
