--- examples/layer_blending/index.html
@@ -1,57 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Layer Blending</title>
-    <style>
-      html,
-      body {
-        margin: 0;
-        padding: 0;
-        display: flex;
-        width: 100%;
-        height: 100%;
-      }
-
-      #main {
-        display: flex;
-        flex-direction: column;
-        width: 100%;
-        height: 100%;
-      }
-
-      #canvas {
-        flex: 1;
-        min-height: 0;
-      }
-
-      #slider {
-        margin: 1rem 2rem;
-      }
-    </style>
-  </head>
-
-  <body>
-    <div id="main">
-      <canvas id="canvas"></canvas>
-      <input type="range" id="slider" />
-      <select id="blendMode">
-        <option value="normal">Normal</option>
-        <option value="additive">Additive</option>
-        <option value="multiply">Multiply</option>
-        <option value="subtractive">Subtractive</option>
-      </select>
-      <label for="opacitySlider">Overlay Opacity</label>
-      <input
-        type="range"
-        id="opacitySlider"
-        min="0"
-        max="1"
-        step="0.01"
-        value="0.5"
-      />
-    </div>
-    <script type="module" src="./main.ts"></script>
-  </body>
-</html>
