--- examples/projected_lines/index.html
@@ -1,26 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Projected Lines</title>
-    <style>
-      html,
-      body {
-        margin: 0;
-        padding: 0;
-        display: flex;
-        width: 100%;
-        height: 100%;
-      }
-      canvas {
-        width: 100%;
-        height: 100%;
-      }
-    </style>
-  </head>
-  <body>
-    <canvas id="canvas"></canvas>
-    <script type="module" src="./main.ts"></script>
-  </body>
-</html>
