+++ src/index.ts
@@ -21,16 +21,13 @@ export {
 export type { ViewportConfig } from "./core/viewport";
 
 export { AxesLayer } from "./layers/axes_layer";
-export { ProjectedLineLayer } from "./layers/projected_line_layer";
-export { TracksLayer } from "./layers/tracks_layer";
-export { ChunkedImageLayer } from "./layers/chunked_image_layer";
-export { VolumeLayer } from "./layers/volume_layer";
 export { ImageLayer } from "./layers/image_layer";
+export type { ImageLayerProps } from "./layers/image_layer";
+export { VolumeLayer } from "./layers/volume_layer";
 export type { Chunk, ChunkLoader, SliceCoordinates } from "./data/chunk";
 export type { QueueStats } from "./core/chunk_manager";
 export { LabelImageLayer } from "./layers/label_image_layer";
 export type { PointPickingResult } from "./layers/point_picking";
-export { ImageSeriesLayer } from "./layers/image_series_layer";
 export { OmeZarrImageSource } from "./data/ome_zarr/image_source";
 
 export type {
