+++ test/renderable_pool.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test, vi, beforeEach } from "vitest";
 import { RenderablePool } from "@/utilities/renderable_pool";
 import { RenderableObject } from "@/core/renderable_object";
-import { poolKeyForImageRenderable } from "@/layers/chunked_image_layer";
+import { poolKeyForImageRenderable } from "@/layers/image_layer";
 import { makeChunk } from "./helpers";
 
 class RenderableStub extends RenderableObject {
